// src/features/types.ts
export type Flight = {
  origin: string;
  origin_name: string;
  destination: string;
  destination_name: string;
  departure_date: string;
  departure_time: string;
  arrival_date: string;
  arrival_time: string;
  carrier: string;
  stops: number;
  price: number;
};

export enum Stops {
  ALL = "Все",
  DIRECT = "Без пересадок",
  ONE = "1 пересадка",
  TWO = "2 пересадки",
  THREE = "3 пересадки",
}